Add files via upload
4x4 tic-tac-toe is played on a board with four rows (numbered 0 to 3 from top to bottom) and 
four columns (numbered 0 to 3 from left to right). There are two players, x and o, who move 
alternately with x always going first. The game is won by the first player to get four of their 
pieces on the same row, column, or diagonal. If the board is full and neither player has won 
then the game is a draw. 
Assuming that it is x's turn to move, x is said to have a forced win if x can make a move 
such that no matter what moves o makes for the rest of the game, x can win. This does not 
necessarily mean that x will win on the very next move, although that is a possibility. It means 
that x has a winning strategy that will guarantee an eventual victory regardless of 
what o does. 
Your task is, given a (possibly) partially-completed game with x to move next, to determine 
whether x has a forced win. If there is such a forced win, the first such forced win found 
should be reported. For this problem, the first forced win is determined by board position, not 
the number of moves required for victory. Search for a forced win by examining positions (0, 
0), (0, 1), (0, 2), (0, 3), (1, 0), (1, 1), ..., (3, 2), (3, 3), in that order, and output the first forced win 
you find. Note that there could be many moves for x that could cause a forced win - you must 
find the first such forced win in this manner.